% Generated by build script
\name{lei_relationships}
\alias{lei_relationships}
\title{Get parent/child relationships for a LEI entity}
\usage{
lei_relationships()
}
\arguments{
\item{lei}{The 20-character LEI code}
}
\value{tibble: relationship_type, start_lei, start_name, end_lei, end_name, status}
\description{Get parent/child relationships for a LEI entity}
